+++ lib/shrine/plugins/mirroring.rb
@@ -3,11 +3,8 @@
 class Shrine
   module Plugins
     module Mirroring
-      UPLOAD = -> (uploaded_file) { uploaded_file.mirror_upload }
-      DELETE = -> (uploaded_file) { uploaded_file.mirror_delete }
-
       def self.configure(uploader, **opts)
-        uploader.opts[:mirroring] ||= { upload: UPLOAD, delete: DELETE }
+        uploader.opts[:mirroring] ||= { upload: true, delete: true }
         uploader.opts[:mirroring].merge!(opts)
 
         fail Error, ":mirror option is required for mirroring plugin" unless uploader.opts[:mirroring][:mirror]
@@ -26,36 +23,61 @@ class Shrine
           end
         end
 
-        def mirror_upload(&block)
+        def mirror_upload_block(&block)
           if block
-            opts[:mirroring][:upload] = block
+            opts[:mirroring][:upload_block] = block
           else
-            opts[:mirroring][:upload]
+            opts[:mirroring][:upload_block]
           end
         end
 
-        def mirror_delete(&block)
+        def mirror_delete_block(&block)
           if block
-            opts[:mirroring][:delete] = block
+            opts[:mirroring][:delete_block] = block
           else
-            opts[:mirroring][:delete]
+            opts[:mirroring][:delete_block]
           end
+        end
+
+        def mirror_upload?
+          opts[:mirroring][:upload]
+        end
+
+        def mirror_delete?
+          opts[:mirroring][:delete]
         end
       end
 
       module InstanceMethods
-        def upload(io, **options)
-          uploaded_file = super
-
-          if self.class.mirrors[storage_key] && self.class.mirror_upload
-            self.class.mirror_upload.call(uploaded_file)
-          end
-
-          uploaded_file
+        # Mirrors upload to other mirror storages.
+        def upload(io, mirror: true, **options)
+          file = super(io, **options)
+          file.trigger_mirror_upload if mirror
+          file
         end
       end
 
       module FileMethods
+        # Mirrors upload if mirrors are defined. Calls mirror block if
+        # registered, otherwise mirrors synchronously.
+        def trigger_mirror_upload
+          return unless shrine_class.mirrors[storage_key] && shrine_class.mirror_upload?
+
+          if shrine_class.mirror_upload_block
+            mirror_upload_background
+          else
+            mirror_upload
+          end
+        end
+
+        # Calls mirror upload block.
+        def mirror_upload_background
+          fail Error, "mirror upload block is not registered" unless shrine_class.mirror_upload_block
+
+          shrine_class.mirror_upload_block.call(self)
+        end
+
+        # Uploads the file to each mirror storage.
         def mirror_upload
           previously_opened = opened?
 
@@ -71,16 +93,33 @@ class Shrine
           end
         end
 
-        def delete
-          result = super
-
-          if shrine_class.mirrors[storage_key] && shrine_class.mirror_delete
-            shrine_class.mirror_delete.call(self)
-          end
-
+        # Mirrors delete to other mirror storages.
+        def delete(mirror: true)
+          result = super()
+          trigger_mirror_delete if mirror
           result
         end
 
+        # Mirrors delete if mirrors are defined. Calls mirror block if
+        # registered, otherwise mirrors synchronously.
+        def trigger_mirror_delete
+          return unless shrine_class.mirrors[storage_key] && shrine_class.mirror_delete?
+
+          if shrine_class.mirror_delete_block
+            mirror_delete_background
+          else
+            mirror_delete
+          end
+        end
+
+        # Calls mirror delete block.
+        def mirror_delete_background
+          fail Error, "mirror delete block is not registered" unless shrine_class.mirror_delete_block
+
+          shrine_class.mirror_delete_block.call(self)
+        end
+
+        # Deletes the file from each mirror storage.
         def mirror_delete
           each_mirror do |mirror|
             self.class.new(id: id, storage: mirror).delete
@@ -89,6 +128,7 @@ class Shrine
 
         private
 
+        # Iterates over mirror storages.
         def each_mirror(&block)
           mirrors = shrine_class.mirrors(storage_key)
           mirrors.map(&block)